serverside/requirements.txt
websockets
python-dotenv
pyserial
numpy